+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Option to upgrade all packages in a scope (eg @capacitor)
 - Migration for Cordova projects using @ionic/angular-toolkit >= v6 (adds @ionic/cordova-builders)
 - Migration for Capacitor projects using @ionic/angular-toolkit >= v6 (cleans angular.json)
+- Detect projects where minifyEnabled is set to true and @capacitor/android is less than 3.2.3
 - Fix for Capacitor commands when @capacitor/cli is not installed
 - Fix for malformed node_module package.json files